python_app/requirements.txt
tkinter
matplotlib
pandas
numpy
scikit-learn
pyserial
sqlite3